shift_bot/requirements.txt
# Telegram shift report bot — MVP
# Python 3.10+

aiogram==3.13.1
aiohttp>=3.9.0
python-dotenv==1.0.1
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
